pyproject.toml
[tool.poetry]
name = "firewhale"
version = "0.1.0"
description = ""
authors = ["Ethan Knapp <ethan@matchlighter.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
docker = "^7.1.0"
ansibleguy-nftables = "^1.0.2.post1"
pyyaml = "^6.0.2"
setuptools = "^75.2.0"
typer = "^0.12.5"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
